refactor: moved map to features, change index.tsx for pages to pages.tsx
