+++ client/src/app/app.tsx
@@ -4,19 +4,18 @@ import './i18n';
 import { Suspense } from 'react';
 import { BrowserRouter } from 'react-router-dom';
 import { Router } from './router';
-import { pages } from 'pages';
 import { Layout } from 'widgets/layout';
 import { GlobalSpin } from 'shared/ui/global-spin';
+import { pages } from 'pages';
 
 export const App = () => {
   return (
     <BrowserRouter>
       <Layout>
-        <Suspense fallback={<GlobalSpin size='large' />}>
+        <Suspense fallback={<GlobalSpin size="large" />}>
           <Router pages={pages} />
         </Suspense>
       </Layout>
     </BrowserRouter>
   );
 };
-
