+++ client/src/features/map/model/map.types.ts
